+++ skoolkit/skoolhtml.py
@@ -1237,6 +1237,9 @@ class HtmlWriter:
     def expand_udgtable(self, text, index, cwd):
         return self.expand_table(text, index, cwd)
 
+    def expand_version(self, text, index, cwd):
+        return skoolmacro.parse_version(text, index)
+
     def expand(self, text, cwd=None):
         """Return `text` with skool macros expanded. `cwd` is the current
         working directory, which is required by macros that create images or